pyproject.toml
[project]
name = "cxg-census-mcp"
version = "0.3.0"
description = "Community Model Context Protocol (MCP) server for the CZ CELLxGENE Discover Census single-cell atlas: ontology-aware filters, provenance, caps, export_snippet. Independent project, not affiliated with CZI, EBI, or any government statistical agency."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Max M. Lang" }]
maintainers = [{ name = "Max M. Lang" }]
keywords = [
    "mcp",
    "model-context-protocol",
    "single-cell",
    "single-cell-rna-seq",
    "scrna-seq",
    "cellxgene",
    "cellxgene-census",
    "ontology",
    "cell-ontology",
    "uberon",
    "mondo",
    "efo",
    "bioinformatics",
    "computational-biology",
    "llm",
    "agents",
    "claude",
    "cursor",
    "anthropic",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries",
    "Operating System :: OS Independent",
]

dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "anyio>=4.3",
    "rapidfuzz>=3.6",
    "diskcache>=5.6",
    "pyarrow>=15.0",
    "structlog>=24.1",
    "platformdirs>=4.2",
    "packaging>=24.0",
]

[project.optional-dependencies]
census = [
    "cellxgene-census>=1.15",
    "tiledbsoma>=1.11",
]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.20",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pyyaml>=6.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
cxg-census-mcp = "cxg_census_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/MaxMLang/cxg-census-mcp"
Issues = "https://github.com/MaxMLang/cxg-census-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cxg_census_mcp"]

[tool.hatch.build.targets.sdist]
include = ["src/", "README.md", "LICENSE", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle / pyflakes
    "I",                # isort
    "UP",               # pyupgrade
    "B",                # bugbear
    "SIM",              # simplify
    "C4",               # comprehensions
    "RUF",              # ruff-specific
]
ignore = [
    "E501",             # line length handled by formatter
    "B008",             # function-call default args (pydantic Field)
    "SIM108",           # ternary if-else
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "SIM117"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live: hits real OLS / Census endpoints (skipped in CI)",
    "slow: long-running tests",
]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]

[tool.coverage.run]
source = ["src/cxg_census_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
